frame_capture/requirements.txt
mss
Pillow
python-dotenv
requests
transformers>=5.0.0
torch
pypdfium2
